refactor(frontend): decompose AssignmentEditor, dedupe create/edit form
AssignmentEditor.tsx previously had three components stuffed into one
file (506 lines) with the create form and the edit form duplicated
verbatim. Split into:

- components/assignment/AssignmentEditor.tsx   — orchestrator (150 lines)
- components/assignment/editor/AssignmentForm.tsx    — shared form used
  for both "create" and "edit" (single source of truth)
- components/assignment/editor/AssignmentItem.tsx    — expandable row
  with inline edit + lazy-loaded submissions
- components/assignment/editor/SubmissionGrader.tsx  — per-submission
  grading card
- components/assignment/editor/types.ts        — AssignmentFormState +
  helpers (assignmentToFormState, formStateToPayload) that normalize
  the DB shape ↔ UI shape conversion
- components/assignment/editor/index.ts        — barrel

No behavior changes. Keyboard-toggle expand uses `void` on the async
handler to satisfy lint. Load effect converted to a promise chain.

tsc, eslint, vitest (85/85), build all pass.

Made-with: Cursor
