+++ frontend/src/components/assignment/editor/index.ts
@@ -1,0 +1,9 @@
+export { AssignmentForm } from "./AssignmentForm"
+export { AssignmentItem } from "./AssignmentItem"
+export { SubmissionGrader } from "./SubmissionGrader"
+export {
+  assignmentToFormState,
+  EMPTY_ASSIGNMENT_FORM,
+  formStateToPayload,
+  type AssignmentFormState,
+} from "./types"
